+++ modules/Deploy/deploy.js
@@ -1,6 +1,7 @@
 const tezsterManager = require("../../tezster-manager");
+const deployContractManager = require("./deploy-contract-manager");
 
-async function deployContract() {
+async function __deployContract() {
 var args = process.argv.slice(3);
     if (args.length < 4) {
         console.log(tezsterManager.outputInfo("Incorrect usage of deploy command \nCorrect usage: - tezster deploy <contract-label> <contract-absolute-path> <init-storage-value> <account>"));
@@ -8,9 +9,9 @@ var args = process.argv.slice(3);
     }
     await tezsterManager.loadTezsterConfig(); 
 
-    let result = await tezsterManager.deployContract(args[0], args[1], args[2], args[3]);
+    let result = await deployContractManager.deployContract(args[0], args[1], args[2], args[3]);
     console.log(result);
     console.log(tezsterManager.outputInfo(`If you're using babylonnet node, use https://babylonnet.tzstats.com to check contract/transactions`));
 }
 
-module.exports = { deployContract };
+module.exports = { __deployContract };
