+++ testcontract.tz
@@ -1,0 +1,3 @@
+parameter string;
+storage string;
+code {CAR; NIL operation; PAIR;};
